requirements.txt
Flask==2.0.2
gunicorn
numpy==1.24.2
openai==0.27.0
pandas==1.5.3
matplotlib==3.7.1
plotly==5.13.1
scipy==1.10.1
scikit-learn==1.2.1
azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
